// src/app/services/model/shipping.ts
export class Shipping {
    id: number;
    name: string;
    companyname: string;
    email: string;
    phone: string;
    country: string;
    streetaddress: string;
    apartmentaddress: string;
    city: string;
    state: string;
    postcode: string;
    userid: number;
    created_at:string;
    updated_at: string
}